backend/requirements.txt
django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
whitenoise==6.9.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
